moved MockRoleVerifier
